refactor: 긴 DAG 코드 리팩토링
 - 일단 다음과 같이 정리
 - 공통으로 쓰는것 - common
 - 각자 쓰는것 - utils
